Add charset helpers
The charset helper takes some string as input, tries to detect
the encoding.

* charsethelper.detect returns the encoding
* charsethelper.decode returns the decoded string without BOM

The charsethelper works similar to fast-chardet, but still relies on
chardet in case there is in fact an uncommon encoding. However, as utf-8
is in fact most common (most translations originate from Babelzilla,
which uses utf-8 sans BOM per default) and ascii is a subset of utf-8
the vast majority of input strings will never hit chardet.